base: promote buffer management into the core, fix gap-buffer bugs
The single-buffer design forced patches to re-home view state and undo,
which 01-buffers did with a fragile shallow-copy mirror that silently
corrupted content on switch (buf_sync copied only view fields while the
active copy and the slot shared gb.data). Fold the whole buffer list and
per-buffer view state into the base so patches build on a stable core:

- Buffer now carries view state (cur/top/left/mark/dirty); BufList +
  bl_init/bl_add/bl_remove/bl_switch live in buffer.c
- Editor.bl; buf_sync (full struct copy) / buf_load (lifts saved view,
  resets undo) / buf_add / buf_switch in edit.c
- edit_init/edit_open create/switch buffers; :bn/:bp/:ls, H/L, status
  bar shows n/m
- quit checks every buffer for unsaved changes (:q/:q!/:wq/Ctrl-Q)

Also fix real bugs found while testing:
- gb_grow checked content size instead of gap size, so typing ~512 chars
  at one spot wrote past the heap allocation (ASan-confirmed)
- gb_rfind returned -1 when before > len, breaking search-prev wrap (N)
- buf_line_of underflowed on a buffer without a built line table
- buf_open accepted directories/FIFOs as editable files
- ESC + printable key decoded as Alt-key, so ESC then a quick key in
  insert mode inserted a literal char instead of exiting insert
- count '0' digit could grow ncount unbounded; paste-before left the
  cursor behind; terminal flags signedness; fuzzy.c param shadow

Verified with ASan/UBSan gap-buffer differential tests and PTY-driven
end-to-end tests (overflow, ESC+j, undo, search wrap, buffer switches).

+++ main.c
@@ -205,7 +205,8 @@ static void render_term_row(struct abuf *ab, int y) {
     int curfg = -1, curbg = -1, curbold = -1;
     for (int x = 0; x < ncols; x++) {
         Cell *c = NULL;
-        if (x < term.cols) c = &term.cells[(size_t)y * term.cols + x];
+        if (x < term.cols)
+            c = &term.cells[(size_t)y * (size_t)term.cols + (size_t)x];
         char ch = c ? (char)c->ch : ' ';
         int fg = c ? c->fg : -1;
         int bg = c ? c->bg : -1;
